requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
line-bot-sdk==1.19.0
MarkupSafe==1.1.1
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.12
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
